fix(featured): fix link on column image generated from CTA
